Change formatting
I changed the fontsize to use 'pt' because just the number wasn't
working. Also added line margins.

I also changed the document class to a KOMAscript book, which
incidentally uses chapters as the top-level-division, so I don't have to
write that one manually. There are pros and cons to different document
classes, but a scrbook appears to be working for me at this time.