Finish consonant alternation section
I removed n~ng alternations from there because there's not much data and
what is there is messy. It's not that important anyway, and I can just
add a paragraph to phonological descriptions instead.

The rest of the consonant alternations are done, although I may need
another section on things like musaʔ~usalan or mabaiy~binasun. Need to
think about that.