Add appendix
I tried to be fancy and import it as a CSV file into a LaTeX table
programmatically, but that turned out to be a pain in the ass, because
neither of the two packages I tried worked well with the IPA and the
quoting and whatnot. It was easier to just take the CSV and manually
edit it into a LaTeX table. That's why the table file itself is in .tex
format.

The two packages in the metadata file are needed for the table. One to
have a table with line breaks (longtable), and the other to have the
pages in landscape mode (pdflscape).

After this I just use \include{} in the main file.