Add highlighting to Chapter 6
I'm using a macro for highlighting, so it'll be easy to change the
command in the future. Right now the colorbox I'm using is pretty small,
and doesn't fit 2 characters well, but is really snug for a single char.

I should make a second macro for 2 chars, or else just use colorbox with
defaults (although the margins are massif on this one).