Add Li's (1985) criteria for Squliq/C'uli'
This is probably the best place for it, since I'm discussing the history
of Atayal subgrouping and presenting my research questions.

I wasn't sure about the last sentence in the paragraph: mentioning my
findings and the NNA subgroup, but it if I omit this, then the
motivation for my dissertation looks weaker.