Add bimoraic vowel analysis for coalescence
Possible analysis of vowel coalescence as a synchronic phonological
phenomenon. I think this works better for verbal suffixation than for
words with vowel coalescence in the root, and wrote as such.

I added descriptions in two places: the vowel coalescence section in the
dialect phonology chapter, and in the PA to Squliq sound changes section
in the PA phonology chapter.