Change List of Abbreviations header
The previous version had it as a section (Level 2) header. I wanted it
as a chapter (Level 1 header), to be consistent with the list of tables
and list of figures, but:

- It still had to show up as section-level in the ToC, again for
consistency
- Chapter headings tend to insert a blank page before the start of the
chapter, using either a \clearpage (in papers) or \cleardoublepage (in
books) command. I did not want this blank page.
- The heading itself had to be unnumbered.

In order to achieve this, I combined the following:

1. \addcontentsline to manually add it to the ToC with the correct name
and heading level. This has to go AFTER the heading itself to be
hyperlinked properly in the PDF.

2. Instead of changing the internals of the \chapter command or trying
to match its output manually, you can supress its aspects individually.
That's what the `\let\cleardoublepage\relax` line is for. In order for
this relaxation to affect only a single chapter instead of all the
chapters in the document, it needs to be wrapped up in a `\begingroup
\endgroup` clause. Curly brackets are supposed to work the same way, but
they didn't for me; this is probably related to Pandoc's parsing of
inlined LaTeX.

I found the supression command on https://tex.stackexchange.com/a/474328

3. An asterisk after a section command (e.g. \chapter* or \section*)
supresses both the numbering and its appearance in the ToC. The Pandoc
Markdown `.unnumbered` class only supresses numbering, but not the ToC
entry, just like the \addchap and \addsec commands (see
https://en.wikibooks.org/wiki/LaTeX/Document_Structure#Section_numbering).
I needed to supress both, because I was manually adding a different
level heading to the ToC.