Disambiguate author names in cites
This one took me a while. I'm citing papers by different Huangs, Chens,
Lius, Chengs, etc., so obviously I need to disambiguate the author
names. Adding an initial is standard procedure. I tried for days to find
the correct option in the CSL manual, and was perplexed when it wasn't
working the way I expected it to.

Turns out there's a bug in pandoc-citeproc that affects this very issue:
https://github.com/jgm/pandoc-citeproc/issues/408

I took the easy way out and added the initials manually, but this is
certainly something that needs to be fixed upstream.