Do not justify text in the appendix table
The 'p' column command that I'm using here to allow multiple columns in
a single table cell also fully justifies the text, with I don't want.

To cancel this justification, I need to pass the >{\raggedright} option
in the column list of the table right before the 'p', as per:
https://tex.stackexchange.com/questions/387642/justify-text-in-a-column

Unfortunately, this redefines the \\ command in tables and makes the
result uncompilable, so an extra \arraybackslash has to be passed
together with \raggedright, as explained here:
https://tex.stackexchange.com/questions/35348/errors-with-ragged-right-justification-in-tables

If you're manually breaking lines in the table, make sure to use
\newline instead of \\.

The \arraybackslash command also requires the array package, added in
metadata.yaml.