Do not align paragraphs to page bottom
In some cases where I have a table/heading that is moved to the next
page, the preceding page is not fully covered in text. By default,
KOMAScript classes will vertically justify paragraphs, but an
unfortunate side-effect is that this creates huge gaps between the
paragraphs themselves. It's better not to justify vertically and leave a
single gap at the bottom in these situations.