Implement minor formatting changes
Have the abstract as a level 1 heading, but keep it out of the ToC.

In a couple of places, pandoc/LaTeX bug out and have a table dip way
down into the bottom margin. I added \newpage commands for these. Sure,
this is antithetical to the spirit of LaTeX, but LaTeX should know
better than to stuff a table and following text into the margin and all
the way off the page.