Format chapter headings
This does two things:

- Centers the chapter headings
- Prepends "Chapter N" on a separate line

The line numbers only get prepended to chapters that have a number.

This tiny fix took me a lot of trial and error, because there are way
too many options in KOMA-Script, and it's confusing having to wade
through them.

I initially tried the `titlesec` package, but couldn't get it to work
with Pandoc. Luckily KOMA-Script provides its own ways of formatting
chapter headings, but you need to wade through its hefty manual to get
to them.