Transform most Markdown tables into LaTeX tables
Pandoc's tables get compiled into LaTeX \longtable by default, which is
a problem if I expressly do not want tables to be broken along a page
break. I could not find any provision in pandoc or a filter that
addresses this problem (I should write one), so I wrote an Elisp
function to automate the conversion process. I had to choose which
tables to convert, because some do still need to be broken up, but it
was a lot easier and faster this way.